# contract/openapi.yaml
openapi: 3.1.0
info:
  title: AI 健身助手 P0 API
  version: 1.0.0
  description: >-
    微信首发但渠道无关的业务契约。关键训练数字由版本化确定性规则产生，
    AI 仅提供候选、排序和解释；用户归属始终由认证主体确定。
servers:
  - url: /
    description: local/test relative base
security:
  - bearerAuth: []
...
paths:
  /api/v1/auth/wechat/login:
    post:
      tags: [Auth]
      operationId: loginWithWechatCode
      security: []
      requestBody: {$ref: '#/components/requestBodies/WechatLogin'}
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '503': {$ref: '#/components/responses/DefaultError'}
        default: {$ref: '#/components/responses/DefaultError'}
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshSession
      security: []
      requestBody: {$ref: '#/components/requestBodies/RefreshSession'}
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        default: {$ref: '#/components/responses/DefaultError'}
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        default: {$ref: '#/components/responses/DefaultError'}
...
  /api/v1/plans/candidates:
    post:
      tags: [Plans]
      operationId: generatePlanCandidates
      description: 接收小程序 AI 生成的结构化计划并由服务端规则校验；AI 不可用或一次修复仍失败时，可显式请求规则保底计划。
      requestBody: {$ref: '#/components/requestBodies/PlanCandidates'}
      responses:
        '200':
          description: 已校验的 AI 个性化候选、明确标记的规则保底候选，或可操作失败原因
          content:
            application/json:
              schema: {$ref: './schemas/plan.yaml#/components/schemas/PlanCandidateGenerationResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/VersionConflict'}
        default: {$ref: '#/components/responses/DefaultError'}
  /api/v1/plans/generation-context:
    get:
      tags: [Plans]
      operationId: getPlanGenerationContext
      description: 返回服务端规范化用户资料、当前可用动作白名单及规则硬约束，供小程序 AI 生成候选；不包含凭据或服务端内部提示。
      parameters:
        - name: profileVersion
          in: query
          required: true
          schema: {$ref: './schemas/common.yaml#/components/schemas/ExpectedVersion'}
      responses:
        '200':
          description: 当前用户的计划生成事实与约束
          content:
            application/json:
              schema: {$ref: './schemas/plan.yaml#/components/schemas/PlanGenerationContextResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/VersionConflict'}
        default: {$ref: '#/components/responses/DefaultError'}
...
  /api/v1/plans/validate:
    post:
      tags: [Plans]
      operationId: validatePlanDraft
      requestBody: {$ref: '#/components/requestBodies/PlanValidation'}
      responses:
        '200':
          description: 按指定规则版本校验计划草稿
          content:
            application/json:
              schema: {$ref: './schemas/plan.yaml#/components/schemas/PlanValidationResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        default: {$ref: '#/components/responses/DefaultError'}
  /api/v1/plans:
    post:
      tags: [Plans]
      operationId: createInitialPlan
      requestBody: {$ref: '#/components/requestBodies/PlanCreation'}
      responses:
        '201':
          description: 已激活候选计划；首次创建计划，已有计划时追加不可变的新版本
          content:
            application/json:
              schema: {$ref: './schemas/plan.yaml#/components/schemas/ActivePlanResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        default: {$ref: '#/components/responses/DefaultError'}
...
components:
  schemas:
...
    ExpectedVersion: {$ref: './schemas/common.yaml#/components/schemas/ExpectedVersion'}
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  requestBodies:
    WechatLogin: {required: true, content: {application/json: {schema: {$ref: './schemas/profile.yaml#/components/schemas/WechatLoginRequest'}}}}
    RefreshSession: {required: true, content: {application/json: {schema: {$ref: './schemas/profile.yaml#/components/schemas/RefreshSessionRequest'}}}}
...
    PlanCandidates: {required: true, content: {application/json: {schema: {$ref: './schemas/plan.yaml#/components/schemas/PlanCandidateRequest'}}}}
    PlanValidation: {required: true, content: {application/json: {schema: {$ref: './schemas/plan.yaml#/components/schemas/ValidatePlanRequest'}}}}
    PlanCreation: {required: true, content: {application/json: {schema: {$ref: './schemas/plan.yaml#/components/schemas/CreatePlanRequest'}}}}
...
  responses:
    Success: {$ref: './schemas/common.yaml#/components/responses/Success'}
    BadRequest: {$ref: './schemas/common.yaml#/components/responses/BadRequest'}
    Unauthorized: {$ref: './schemas/common.yaml#/components/responses/Unauthorized'}
...
    VersionConflict: {$ref: './schemas/common.yaml#/components/responses/VersionConflict'}
    TooManyRequests: {$ref: './schemas/common.yaml#/components/responses/TooManyRequests'}
    DefaultError: {$ref: './schemas/common.yaml#/components/responses/DefaultError'}